Rename incorrect variable name 1. Quesiton -> Question 2. categories -> topics
