+++ frontend/src/pages/BankPage/Bank.page.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { QnTable } from "../../components/QnTable/QnTable.component";
 import { FilterBar } from "../../components/QnFilter/QnFilter.component";
 import { VStack } from "@chakra-ui/react";
-import { QnFilter } from "../../models/Quesiton.model";
-import { Question } from "../../models/Quesiton.model";
+import { QnFilter } from "../../models/Question.model";
+import { Question } from "../../models/Question.model";
 import { dummyQn } from "../../data/sampleqn";
 import { useDispatch, useSelector } from "react-redux";
 import { selectFilteredQuestions, setQuestions } from "../../reducers/questionsSlice";
