+++ frontend/src/pages/HomePage/HomePage.tsx
@@ -3,7 +3,7 @@ import "./HomePage.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { useNavigate } from "react-router-dom";
 import { Question, initializeData } from "../AddQuestionPage/QuestionData";
-import { MarkdownEditor } from "../../components/MarkdownEditor/MarkdownEditor.component";
+import { QuestionEditor } from "../../components/QuestionEditor/QuestionEditor.component";
 
 function HomePage() {
   let history = useNavigate();
@@ -76,7 +76,7 @@ function HomePage() {
     setFilteredData(filtered);
   };
 
-  return <MarkdownEditor></MarkdownEditor>;
+  return <QuestionEditor></QuestionEditor>;
 }
 
 export default HomePage;
